+++ src/components/QuestionsAnswers/MasterQAContainer/Question.jsx
@@ -1,14 +1,18 @@
 import React from "react";
 import Answers from "./Answers";
 import Helpful from "./Helpful";
+import Axios from "axios";
 import { render } from "@testing-library/react";
 
 class Question extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      style: null
+      style: null,
+      answers: []
     };
+    this.setAnswers = this.setAnswers.bind(this);
+    this.getAnswers = this.getAnswers.bind(this);
   }
 
   expandComponent() {
@@ -18,6 +22,27 @@ class Question extends React.Component {
       overflowY: "scroll"
     };
     this.setState({ style: newStyles });
+  }
+
+  getAnswers(id) {
+    return Axios.get(`http://3.134.102.30/qa/${id}/answers`);
+  }
+
+  setAnswers() {
+    this.getAnswers(this.props.q.question_id).then(res => {
+      let sortedAnswers = this.sortAnswers(res.data.results);
+      this.setState({ answers: sortedAnswers }, () => console.log(this.state));
+    });
+  }
+
+  componentDidMount() {
+    this.setAnswers();
+  }
+
+  sortAnswers(array) {
+    let answers = array.slice(0);
+    answers.sort((a, b) => (a.helpfulness > b.helpfulness ? -1 : 1));
+    return answers;
   }
 
   render() {
@@ -39,6 +64,7 @@ class Question extends React.Component {
                   questionID={questionID}
                   helpful={this.props.q.question_helpfulness}
                   type="question"
+                  setAnswers={this.setAnswers}
                 />
               </div>
             </td>
@@ -47,9 +73,9 @@ class Question extends React.Component {
             expandHandler={() => {
               this.expandComponent();
             }}
-            questionID={questionID}
-            // answers={answers}
+            answers={this.state.answers}
             key={questionID}
+            setAnswers={this.setAnswers}
           />
         </tbody>
       </div>
